Struct rusoto_sts::StsClient

pub struct StsClient { /* fields omitted */ }

A client for the AWS STS API.

Methods

impl StsClient

pub fn new(region: Region) -> StsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> StsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,