Struct rusoto_sts::StsSessionCredentialsProvider

pub struct StsSessionCredentialsProvider { /* fields omitted */ }

AwsCredentials provider that calls GetSessionToken using the provided StsClient. To use with MFA, pass in the MFA serial number then set the MFA code. You will need to ensure the provider has a valid code each time you acquire a new STS token.

Methods

impl StsSessionCredentialsProvider

pub fn new(
sts_client: StsClient,
duration: Option<Duration>,
mfa_serial: Option<String>
) -> StsSessionCredentialsProvider

Creates a new StsSessionCredentialsProvider with the given StsClient and session parameters.

sts_client - The StsClient to use to acquire session tokens.

duration - The duration of the session tokens. Default 1 hour.

mfa_serial - Optional MFA hardware device serial number or virtual device ARN. Set the MFA code with set_mfa_code.

pub fn set_mfa_code<S>(&mut self, code: S) where
S: Into<String>,

Set the MFA code for use when acquiring session tokens.

pub fn clear_mfa_code(&mut self)

Clear the MFA code.

pub fn get_session_token(&self) -> StsSessionCredentialsProviderFuture

Calls GetSessionToken to get a session token from the STS Api. Optionally uses MFA if the MFA serial number and code are set.