Enum rusoto_sts::AssumeRoleWithWebIdentityError

pub enum AssumeRoleWithWebIdentityError {
    ExpiredToken(String),
    IDPCommunicationError(String),
    IDPRejectedClaim(String),
    InvalidIdentityToken(String),
    MalformedPolicyDocument(String),
    PackedPolicyTooLarge(String),
    RegionDisabled(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    ParseError(String),
    Unknown(BufferedHttpResponse),
}

Errors returned by AssumeRoleWithWebIdentity

Variants
ExpiredToken(String)

The web identity token that was passed is expired or is not valid. Get a new identity token from the identity provider and then retry the request.

IDPCommunicationError(String)

The request could not be fulfilled because the non-AWS identity provider (IDP) that was asked to verify the incoming identity token could not be reached. This is often a transient error caused by network conditions. Retry the request a limited number of times so that you don't exceed the request rate. If the error persists, the non-AWS identity provider might be down or not responding.

IDPRejectedClaim(String)

The identity provider (IdP) reported that authentication failed. This might be because the claim is invalid.

If this error is returned for the AssumeRoleWithWebIdentity operation, it can also mean that the claim has expired or has been explicitly revoked.

InvalidIdentityToken(String)

The web identity token that was passed could not be validated by AWS. Get a new identity token from the identity provider and then retry the request.

MalformedPolicyDocument(String)

The request was rejected because the policy document was malformed. The error message describes the specific error.

PackedPolicyTooLarge(String)

The request was rejected because the policy document was too large. The error message describes how big the policy document is, in packed form, as a percentage of what the API allows.

RegionDisabled(String)

STS is not activated in the requested region for the account that is being asked to generate credentials. The account administrator must use the IAM console to activate STS in that region. For more information, see Activating and Deactivating AWS STS in an AWS Region in the IAM User Guide.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

ParseError(String)

An error occurred parsing the response payload.

Unknown(BufferedHttpResponse)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl AssumeRoleWithWebIdentityError

pub fn from_response(
res: BufferedHttpResponse
) -> AssumeRoleWithWebIdentityError